+++ ecosystem/typescript/sdk/src/transaction_builder/aptos_types/index.ts
@@ -11,5 +11,6 @@ export * from "./identifier";
 export * from "./ed25519";
 export * from "./multi_ed25519";
 export * from "./authentication_key";
+export * from "./rotation_proof_challenge";
 
 export type SigningMessage = Buffer;